Migrate the render logic into TripController
